bug: 高峰期引擎打印的服务内的构建日志触发熔断 #11589 解耦日志打印的服务依赖
