+++ src/backend/ci/core/common/common-log/src/main/kotlin/com/tencent/devops/common/log/event/LogStorageEvent.kt
@@ -25,7 +25,7 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.tencent.devops.log.event
+package com.tencent.devops.common.log.event
 
 import com.tencent.devops.common.log.pojo.message.LogMessageWithLineNo
 import com.tencent.devops.common.event.annotation.Event
