+++ src/backend/ci/core/common/common-log/src/main/kotlin/com/tencent/devops/common/log/meta/Ansi.kt
@@ -25,7 +25,7 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.tencent.devops.log.meta
+package com.tencent.devops.common.log.meta
 
 import java.util.concurrent.Callable
