+++ src/backend/ci/core/common/common-log/src/main/kotlin/com/tencent/devops/common/log/meta/AnsiAttribute.kt
@@ -1,7 +1,7 @@
 /*
  * Tencent is pleased to support the open source community by making BK-CI 蓝鲸持续集成平台 available.
  *
- * Copyright (C)) 2019 THL A29 Limited, a Tencent company.  All rights reserved.
+ * Copyright (C) 2019 THL A29 Limited, a Tencent company.  All rights reserved.
  *
  * BK-CI 蓝鲸持续集成平台 is licensed under the MIT license.
  *
@@ -11,7 +11,7 @@
  * Terms of the MIT License:
  * ---------------------------------------------------
  * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
- * documentation files (the "Software")), to deal in the Software without restriction, including without limitation the
+ * documentation files (the "Software"), to deal in the Software without restriction, including without limitation the
  * rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to
  * permit persons to whom the Software is furnished to do so, subject to the following conditions:
  *
@@ -25,16 +25,25 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-dependencies {
-    api(project(":core:dispatch:api-dispatch-kubernetes"))
-    api(project(":core:dispatch:model-dispatch"))
+package com.tencent.devops.common.log.meta
 
-    api(project(":core:common:common-service"))
-    api(project(":core:common:common-client"))
-    api(project(":core:common:common-redis"))
-    api(project(":core:common:common-db-base"))
-    api(project(":core:common:common-pipeline"))
-    api(project(":core:common:common-auth:common-auth-api"))
-    api(project(":core:common:common-dispatch-sdk"))
-    api(project(":core:common:common-log"))
+enum class AnsiAttribute(val value: Int) {
+    RESET(0),
+    INTENSITY_BOLD(1),
+    INTENSITY_FAINT(2),
+    ITALIC(3),
+    UNDERLINE(4),
+    BLINK_SLOW(5),
+    BLINK_FAST(6),
+    NEGATIVE_ON(7),
+    CONCEAL_ON(8),
+    STRIKETHROUGH_ON(9),
+    UNDERLINE_DOUBLE(21),
+    INTENSITY_BOLD_OFF(22),
+    ITALIC_OFF(23),
+    UNDERLINE_OFF(24),
+    BLINK_OFF(25),
+    NEGATIVE_OFF(27),
+    CONCEAL_OFF(28),
+    STRIKETHROUGH_OFF(29);
 }
