+++ src/backend/ci/core/common/common-log/src/main/kotlin/com/tencent/devops/common/log/meta/AnsiErase.kt
@@ -25,13 +25,10 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.tencent.devops.common.log.constant
+package com.tencent.devops.common.log.meta
 
-object Constants {
-    const val DEFAULT_PRIORITY_NOT_DELETED: Byte = 40
-    const val RESPONSE_ENTITY_MAX_SIZE = 1024 * 1024 * 1024
-    const val SCROLL_MAX_LINES = 5000
-    const val NORMAL_MAX_LINES = 10000
-    const val BULK_BUFFER_SIZE = 1000
-    const val SCROLL_MAX_TIMES = 6
+enum class AnsiErase(val value: Int) {
+    FORWARD(0),
+    BACKWARD(1),
+    ALL(2);
 }
