+++ src/backend/ci/core/common/common-log/src/main/kotlin/com/tencent/devops/common/log/pojo/QueryLogStatus.kt
