+++ src/backend/ci/core/dispatch/biz-dispatch-docker/build.gradle.kts
@@ -35,8 +35,8 @@ dependencies {
     api(project(":core:common:common-client"))
     api(project(":core:common:common-redis"))
     api(project(":core:common:common-db-base"))
+    api(project(":core:common:common-log"))
     api(project(":core:common:common-auth:common-auth-api"))
-    api(project(":core:log:api-log"))
     api(project(":core:buildless:api-buildless"))
     api(project(":core:dispatch:api-dispatch-kubernetes"))
 }
