+++ src/backend/ci/core/dispatch/biz-dispatch/build.gradle.kts
@@ -32,12 +32,12 @@ dependencies {
     api(project(":core:dispatch:model-dispatch"))
     api(project(":core:notify:api-notify"))
     api(project(":core:misc:api-monitoring"))
-    api(project(":core:log:api-log"))
     api(project(":core:common:common-service"))
     api(project(":core:common:common-web"))
     api(project(":core:common:common-client"))
     api(project(":core:common:common-redis"))
     api(project(":core:common:common-db-base"))
+    api(project(":core:common:common-log"))
     api(project(":core:common:common-auth:common-auth-api"))
     api(project(":core:common:common-dispatch-sdk"))
     api(project(":core:common:common-kafka"))
