+++ src/backend/ci/core/log/api-log/build.gradle.kts
@@ -27,8 +27,7 @@
 
 dependencies {
     api(project(":core:common:common-api"))
-    api(project(":core:common:common-event"))
-    api("io.github.resilience4j:resilience4j-circuitbreaker")
+    api(project(":core:common:common-log"))
 }
 
 plugins {
