+++ src/backend/ci/core/log/api-log/src/main/kotlin/com/tencent/devops/log/constant/Constants.kt
@@ -25,25 +25,13 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.tencent.devops.log.meta
+package com.tencent.devops.log.constant
 
-enum class AnsiAttribute(val value: Int) {
-    RESET(0),
-    INTENSITY_BOLD(1),
-    INTENSITY_FAINT(2),
-    ITALIC(3),
-    UNDERLINE(4),
-    BLINK_SLOW(5),
-    BLINK_FAST(6),
-    NEGATIVE_ON(7),
-    CONCEAL_ON(8),
-    STRIKETHROUGH_ON(9),
-    UNDERLINE_DOUBLE(21),
-    INTENSITY_BOLD_OFF(22),
-    ITALIC_OFF(23),
-    UNDERLINE_OFF(24),
-    BLINK_OFF(25),
-    NEGATIVE_OFF(27),
-    CONCEAL_OFF(28),
-    STRIKETHROUGH_OFF(29);
+object Constants {
+    const val DEFAULT_PRIORITY_NOT_DELETED: Byte = 40
+    const val RESPONSE_ENTITY_MAX_SIZE = 1024 * 1024 * 1024
+    const val SCROLL_MAX_LINES = 5000
+    const val NORMAL_MAX_LINES = 10000
+    const val BULK_BUFFER_SIZE = 1000
+    const val SCROLL_MAX_TIMES = 6
 }
