+++ src/backend/ci/core/log/api-log/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-com.tencent.devops.log.configuration.LogPrinterConfiguration
+com.tencent.devops.common.log.configuration.LogPrinterConfiguration
