+++ src/backend/ci/core/misc/biz-image/build.gradle.kts
@@ -35,7 +35,7 @@ dependencies {
     api(project(":core:common:common-service"))
     api(project(":core:common:common-web"))
     api(project(":core:common:common-auth:common-auth-api"))
-    api(project(":core:log:api-log"))
+    api(project(":core:common:common-log"))
     api(project(":core:dispatch:api-dispatch-kubernetes"))
     api("org.apache.commons:commons-compress")
     api("com.github.docker-java:docker-java-core")
