+++ src/backend/ci/core/openapi/biz-openapi/build.gradle.kts
@@ -31,7 +31,6 @@ dependencies {
     api(project(":core:common:common-db"))
     api(project(":core:common:common-pipeline"))
     api(project(":core:common:common-es"))
-    api(project(":core:common:common-stream"))
     api(project(":core:repository:api-repository"))
     api(project(":core:environment:api-environment"))
     api(project(":core:artifactory:api-artifactory"))
