+++ src/backend/ci/core/process/biz-engine/src/test/kotlin/com/tencent/devops/process/engine/control/DependOnControlTest.kt
@@ -38,12 +38,10 @@ import io.mockk.mockk
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-import org.springframework.cloud.stream.function.StreamBridge
 
 class DependOnControlTest {
     private val pipelineContainerService: PipelineContainerService = mockk()
-    private val streamBridge: StreamBridge = mockk()
-    private val buildLogPrinter: BuildLogPrinter = BuildLogPrinter(streamBridge, mockk())
+    private val buildLogPrinter: BuildLogPrinter = BuildLogPrinter(mockk())
     private val dependOnControl = DependOnControl(
         pipelineContainerService = pipelineContainerService,
         buildLogPrinter = buildLogPrinter
