+++ src/backend/ci/core/process/biz-engine/src/test/kotlin/com/tencent/devops/process/engine/control/MutexControlTest.kt
@@ -45,7 +45,7 @@ import org.junit.jupiter.api.Test
 @Suppress("ALL", "UNUSED")
 class MutexControlTest {
 
-    private val buildLogPrinter: BuildLogPrinter = BuildLogPrinter(mockk(), mockk())
+    private val buildLogPrinter: BuildLogPrinter = BuildLogPrinter(mockk())
     private val redisOperation: RedisOperation = RedisOperation(mockk(), mockk(), mockk())
     private val variables: Map<String, String> = mapOf(Pair("var1", "Test"))
     private val buildId: String = "b-12345678901234567890123456789012"
