+++ src/backend/ci/core/process/biz-process/build.gradle.kts
@@ -32,6 +32,7 @@ dependencies {
     api(project(":core:common:common-redis"))
     api(project(":core:common:common-archive"))
     api(project(":core:common:common-websocket"))
+    api(project(":core:common:common-log"))
     api(project(":core:store:api-store"))
     api(project(":core:dispatch:api-dispatch"))
     api(project(":core:project:api-project"))
@@ -41,7 +42,6 @@ dependencies {
     api(project(":core:misc:api-plugin"))
     api(project(":core:notify:api-notify"))
     api(project(":core:process:biz-base"))
-    api(project(":core:log:api-log"))
     api(project(":core:common:common-webhook:biz-common-webhook"))
     api(project(":core:auth:api-auth"))
     api(project(":core:common:common-pipeline-yaml"))
