+++ src/backend/ci/settings.gradle.kts
@@ -69,7 +69,7 @@ include(":core:common:common-archive")
 include(":core:common:common-quality")
 include(":core:common:common-service")
 include(":core:common:common-pipeline")
-include(":core:common:common-stream")
+include(":core:common:common-log")
 include(":core:common:common-expression")
 include(":core:common:common-test")
 include(":core:common:common-auth")
